two_football/requirements.txt
amqp==2.4.2
aniso8601==5.1.0
APScheduler==3.6.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.6
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
decorator==4.4.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
kombu==4.5.0
MarkupSafe==1.1.1
marshmallow==2.19.2
ndg-httpsclient==0.5.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pusher==2.1.3
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
relativetimebuilder==0.2.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.0
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
